+++ OsirisPdvReal/Models/FornecedorBanca.cs
@@ -7,7 +7,7 @@ namespace OsirisPdvReal.Models
 {
     public class FornecedorBanca
     {
-        public int CNPJ { get; set; }
+        public long CNPJ { get; set; }
         public Fornecedor Fornecedores { get; set; }
         public int BancaId  { get; set; }
         public Banca Bancas { get; set; }
